refactor: give ToastMessage a single serialization format
ToastMessage carried two near-identical array shapes differing only in the
type/variant key. The variant rename only exists because Effect spreads its
payload next to its own type discriminator, so that concern now lives in
Effect::toast. ToastMessage keeps one canonical toArray/jsonSerialize and
moves to Core\Values, with ToastType joining the other Core\Enums.
