+++ src/Actions/ActionResult.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Bambamboole\Lattice\Actions;
 
 use Bambamboole\Lattice\Contracts\Effect as EffectContract;
-use Bambamboole\Lattice\Core\Toasts\Enums\ToastType;
-use Bambamboole\Lattice\Core\Toasts\ToastMessage;
+use Bambamboole\Lattice\Core\Enums\ToastType;
+use Bambamboole\Lattice\Core\Values\ToastMessage;
 use JsonSerializable;
 
 final readonly class ActionResult implements JsonSerializable
