+++ workbench/app/Providers/TypeScriptTransformerServiceProvider.php
@@ -10,8 +10,8 @@ use Bambamboole\Lattice\Core\Enums\Gap;
 use Bambamboole\Lattice\Core\Enums\HttpMethod;
 use Bambamboole\Lattice\Core\Enums\PageContainer;
 use Bambamboole\Lattice\Core\Enums\PageLayout;
+use Bambamboole\Lattice\Core\Enums\ToastType;
 use Bambamboole\Lattice\Core\Enums\Width;
-use Bambamboole\Lattice\Core\Toasts\Enums\ToastType;
 use Bambamboole\Lattice\Forms\Enums\Op;
 use Bambamboole\Lattice\Tables\Enums\PaginationType;
 use Spatie\LaravelTypeScriptTransformer\TypeScriptTransformerApplicationServiceProvider;
